requirements.txt
# Core dependencies for RAG Chatbot
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-google-genai>=1.0.0

# LLM Providers
openai>=1.0.0
google-generativeai>=0.3.0
ollama>=0.1.0  # For Llama models via Ollama

# Embeddings
sentence-transformers>=2.2.0  # For HuggingFace embeddings
transformers>=4.30.0  # Required for HuggingFace models

# Vector Databases
faiss-cpu>=1.7.4
chromadb>=0.4.0

# Utilities
python-dotenv>=1.0.0
streamlit>=1.25.0
llama-index
pypdf
unstructured>=0.10.0
nltk
rank-bm25
pandas
matplotlib
seaborn
tiktoken>=0.5.0
torch>=2.0.0  # Required for sentence-transformers

# Query Preprocessing & Spell Checking
textblob>=0.17.1
pyspellchecker>=0.7.0
rapidfuzz>=3.0.0  # For fuzzy string matching

# RAGAS Evaluation Framework
ragas>=0.1.0
langchain-google-genai>=1.0.0
datasets>=2.14.0
